fix bug:  write variable large than 4G bytes  report " first (xxx) must >=0"
Backup Old ChunkingIndex as ChunkIndexV0.

+++ cdm/core/src/main/java/ucar/nc2/write/NetcdfCopier.java
@@ -368,7 +368,7 @@ public class NetcdfCopier {
           byteWriteTotal += data.getSize();
         }
 
-        index.setCurrentCounter(index.currentElement() + (int) Index.computeSize(chunkShape));
+        index.setCurrentCounter(index.currentElement() + Index.computeSize(chunkShape));
         if (cancel.isCancel()) {
           return;
         }
